fix(Drive): use actual value for initial start drive speed
Using a drive speed that is too high causes problems with Unity joints
and can cause them to break their restrictions.

Instead of using a float max value, it now uses just a high value that
is fast enough to not notice but not fast enough to break the Unity
joints system.

Whilst 5000f seems like a magic number, it's just a number that was
decided that fell into the category of not breaking things but being
fast enough to not notice.