feat(Driver): add methods for properties in UnityEvents
A number of properties cannot be changed via UnityEvents because
they are non-standard data types, so methods have been added that
allow for these data types to be changed.

A small fix for the DriveLimits not being updated has been added
and the Setup code has been refactored so the individual setup
elements can now be run individually if need be.

Also, now when the DriveAxis is changed, it will reset the rotation
on the axis as otherwise the rotations cause adverse effects. It
needs looking into further in the future to see if the rotational
data can be kept when switching axes as this would be a nice feature.