fix(Drive): rework the drive reset functionality
The AngularDrive ResetRotations was causing the MoveToTargetValue
functionality to break because it was always resetting the drive
everytime the target value changed. This has been fixed by reworking
the Reset concept into the core Drive class and having the ability
of only calling ResetDrive the first time SetUp is called and then it
won't call Reset again until a manual intervention occurs.