fix(Drive): expose initial drive value speed as public property
The IntialDriveValueSpeed is now a public changable property rather
than a protected constant field so it can be adapted to suit different
requirements.

The AngularTransformDrive did not work well at such a high speed so
it has now been reduced to make it work correctly using this new
ability to publicly set the initial speed of the drive.