fix(Drive): process the drive correctly after disable and re-enable
There was an issue where the drive would not correctly run the process
code after it was disabled and re-enabled because the process code
relied on the fact that the drive would start in a central position
and move to a different location which would allow the process
mechanism to operate because the value delta has changed.

But upon disabling and re-enabling, the value of the button location
would be the same causing no process to occur which would then in turn
prevent the events from being emitted.

This fix just ensures that upon disable and re-enable the process at
least happens once to ensure the correct set up occurs.