feat(prefab): add tag and string identifiers to drive elements
The drive elements have had a couple of tags and string identifiers
added to the parts of the drive that are reported upon collisions
such as the joint drives will report the joint GameObject and maybe
the Interactable GameObject, whereas the transform drives will only
report the Interactable GameObject.

These additional elements make it easy to identify the drive and
type based on the tag and string of the found colliding object.