fix(Editor): only apply position when auto creating controllable
The Controllable Creator now only applies the original mesh position
to the new controllable prefab and the rotation, scale of the mesh
stay on the original mesh. This is because controllables that have
external rotations and scales will warp the object inside so the
mesh is the thing that needs rotating and scaling.