feat(LinearDriveFacade): editor button to align to initial target value
A new custom editor for the LinearDriveFacade has been added that just
adds a button to the bottom of the component inspector that when
clicked will align the linear drive component in the editor scene to
the position that the initial target value is set to.

This makes it easier to line up the drive object in the editor as it
can be shown at the various stages of its travel.