fix(Facade): ensure gizmos draw correct length with parent scale
There was an issue where the gizmo was not drawing the correct length
if the parent GameObject was scaled as this is affecting the world
scale and therefore the gizmo would be affected also.

As drive limits are world scale, then the gizmo should represent this
by drawing to world scale, which this commit fixes.

The gizmo colour is now also an option on the Facade to allow the color
to be customised if required.