feat(Drive): add IsGrabbable option
The IsGrabbable option on the Drive derivatives allows the toggle of
whether the Interactable Object is grabbable or not, which can easily
simulate locking the drive mechanism.

The Interactable reference is now part of Drive so all inherited drives
can access the InteractabeFacade as well as the MeshContainer if
required.

The Gizmo color option has also been moved to the Drive as it isn't
something that is needed on the Facade and it can then be shared
across all drive types.