fix(AngularDrive): ensure target value threshold is normalized value
The targetValueReachedThreshold property should be a normalized value
as both InitialTargetValue and TargetValue are normalized properties.

But on the AngularDrive it was being treated as the actual rotation
value threshold, which caused confusion as it was different between
drive types and also caused certain events to break as they were
expecting to check off the normalized value.

All prefabs have had their default targetValueReachedThreshold set to
`0.0075` as this is a decent default normalized threshold to consider
the drive is at the target value.