feat(LinearTransformDrive): add reference to artificial velocity
The LinearTransformDrive now has a reference to the Artificial Velocity
Applier component that is used when the Interactable Object is
ungrabbed. This is to keep it inline with the AngularTransformDrive
which also has a reference to the same component.

The ungrabbed/grabbed drag options on both Linear drives has also been
set to 5/5 for consistency.