feat(Drive): add helper events for move to target value
Two new helper events have been added to the internal drive component
that will emit when the Facade.MoveToTargetValue property is changed
so extra processing can be done on that event occurring if needed.

These events will also be processed on enable of the drive so initial
values will be taken into consideration.