feat(Drive): add grabbed target threshold
The Target Value Reached threshold is usually larger for an auto moving
drive so it cannot overshoot, but for a grabbed drive the target
threshold can tend to be too large causing a jitter when the grabbed
drive is moved to the extremes.

This adds a new threshold value for when the drive is grabbed so it
can be even more refined for a better experience.